Rename NServiceBus.Transport.SQS.Tests
