+++ src/AcceptanceTests/AcceptanceTests.csproj
@@ -15,11 +15,11 @@
     <PackageReference Include="NUnit3TestAdapter" Version="3.16.1" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\Tests\Cleanup.cs">
-      <Link>Cleanup.cs</Link>
-    </Compile>
+    <ProjectReference Include="..\NServiceBus.Transport.SQS\NServiceBus.Transport.SQS.csproj" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\NServiceBus.Transport.SQS\NServiceBus.Transport.SQS.csproj" />
+    <Compile Include="..\NServiceBus.Transport.SQS.Tests\Cleanup.cs">
+      <Link>Cleanup.cs</Link>
+    </Compile>
   </ItemGroup>
 </Project>
