+++ src/NServiceBus.Transport.SQS.Tests/MessageDispatcherTests.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.AmazonSQS.Tests
+﻿namespace NServiceBus.Transport.SQS.Tests
 {
     using System;
     using System.Collections;
@@ -8,16 +8,17 @@
     using System.Threading.Tasks;
     using Amazon.SimpleNotificationService.Model;
     using Amazon.SQS.Model;
+    using Configure;
     using DelayedDelivery;
     using DeliveryConstraints;
     using Extensibility;
     using Newtonsoft.Json.Linq;
+    using NServiceBus;
     using NUnit.Framework;
     using Routing;
     using Settings;
+    using SQS;
     using Transport;
-    using Transport.SQS;
-    using Transport.SQS.Configure;
 
     [TestFixture]
     public class MessageDispatcherTests
@@ -202,8 +203,8 @@
             await dispatcher.Dispatch(transportOperations, transportTransaction, context);
 
             Assert.AreEqual(2, mockSnsClient.PublishedEvents.Count);
-            Assert.AreEqual("arn:aws:sns:us-west-2:123456789012:NServiceBus-AmazonSQS-Tests-MessageDispatcherTests-Event", mockSnsClient.PublishedEvents.ElementAt(0).TopicArn);
-            Assert.AreEqual("arn:aws:sns:us-west-2:123456789012:NServiceBus-AmazonSQS-Tests-MessageDispatcherTests-AnotherEvent", mockSnsClient.PublishedEvents.ElementAt(1).TopicArn);
+            Assert.AreEqual("arn:aws:sns:us-west-2:123456789012:NServiceBus-Transport-SQS-Tests-MessageDispatcherTests-Event", mockSnsClient.PublishedEvents.ElementAt(0).TopicArn);
+            Assert.AreEqual("arn:aws:sns:us-west-2:123456789012:NServiceBus-Transport-SQS-Tests-MessageDispatcherTests-AnotherEvent", mockSnsClient.PublishedEvents.ElementAt(1).TopicArn);
         }
 
         [Test]
