+++ src/NServiceBus.Transport.SQS.Tests/NServiceBus.Transport.SQS.Tests.csproj
