+++ src/NServiceBus.Transport.SQS.TransportTests/NServiceBus.Transport.SQS.TransportTests.csproj
@@ -15,13 +15,13 @@
     <PackageReference Include="NUnit3TestAdapter" Version="3.16.1" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\Tests\Cleanup.cs">
-      <Link>Cleanup.cs</Link>
-    </Compile>
     <Compile Include="..\AcceptanceTests\SqsTransportExtensions.cs">
       <Link>SqsTransportExtensions.cs</Link>
     </Compile>
-    <Compile Include="..\Tests\SettingsHolderExtensions.cs">
+    <Compile Include="..\NServiceBus.Transport.SQS.Tests\Cleanup.cs">
+      <Link>Cleanup.cs</Link>
+    </Compile>
+    <Compile Include="..\NServiceBus.Transport.SQS.Tests\SettingsHolderExtensions.cs">
       <Link>SettingsHolderExtensions.cs</Link>
     </Compile>
   </ItemGroup>
